pyproject.toml
[project]
name = "mbx-inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.28.0",
    "mbx-db",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
mbx-inventory = "mbx_inventory.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
